Create QuizApp
The Quiz Application is a Python-based application that allows users to participate in quizzes and provides an administrative interface for managing quiz questions and user reports. The application utilizes the PyMySQL library to interact with a MySQL database for data storage and retrieval. The application consists of two main user roles: admin and user.